fix: deduplicate _fts_search + expose search_vectors_raw (#145, #136)
#145: _fts_search was defined identically in consolidation.py and
personality.py. Moved to fts_search.py (where it logically belongs)
and both files now import from there.

#136: client.py accessed engine._has_vectors and engine.conn directly.
Added search_vectors_raw() public method to TrueMemoryEngine that
encapsulates the vectors-available check and raw cosine search.
Client now calls the public method instead of reaching into internals.
